refactor: consolidate type imports by removing countryGeoJsonTypes and updating references
