+++ src/library/types/index.ts
@@ -1,3 +1,3 @@
 export * from './constantTypes'
 export * from './mapTypes'
-export * from './componentTypes'
+export * from './geoJsonTypes'
